MRIRealign.jl

This package aligns a time series of 3D MRI images with similar contrast, following the seminal paper by Friston et al. It minimizes the squared difference between the images in a given mask. The package was heavily inspired by SPM's spm_realign function. The principal advantage over spm_realign is speed. Additionally, we implemented a consensus estimation, which aligns all time frames pairwise and uses iteratively reweighted least squares to calculate a consensus between all estimates. Compared to a single reference time frame, the consensus approach is less sensitive to the image quality of the reference frame. Compared with the mean time frame as a reference, it avoids difficulties in mapping to a blurred reference.

Quick Tutorial

On Unix systems, Julia can be installed with

curl -fsSL https://install.julialang.org | sh

and on Windows systems with

winget install --name Julia --id 9NJNWW8PVKMN -e -s msstore

More detailed installation instructions can be found here.

Thereafter, you can start Julia from the command line with

julia

This section assumes that you have a folder at the path /path_to_files/ with NIfTI files of the format mask.nii and somename_1.nii, somename_2.nii, ... . Our package does not include loading functions, allowing users to load data from NIfTI, DICOM, Matlab, HDF5 files etc.

The first time, the packages need to be installed with the package manager:

using Pkg
Pkg.add("MRIRealign")
Pkg.add("NIfTI")

Thereafter, we can use them:

using MRIRealign
using NIfTI

We can change the directory

cd("/path_to_files/")

and, optionally, load a mask and convert it to a binary mask

mask = round.(Bool, niread("mask.nii"))

Note that the . after round indicates a point-wise operation.

We can create a list of file names in the current folder, except for mask.nii, and sort them in natural order, i.e., 1, 2, 3, ... instead of the ASCII order 1, 10, 100, 101, ... .

files = filter(f -> isfile(f) && f != "mask.nii", readdir())
files = sort(files, by = file -> parse(Int, match(r"\d+", file).match))

Using the size of the mask, where size(mask)... returns the three dimensions separately, we can allocate an array and load all time frames into it:

img = Array{Float64}(undef, size(mask)..., length(files))

for t in eachindex(files)
    img[:,:,:,t] .= niread(files[t]).raw
end

Now we are all set to call the realign! function, which will overwrite img with the aligned volumes and return the motion parameters, i.e., 3 rotation and 3 translation parameters in this order:

params = realign!(img; mask=mask)

We can write the aligned images back to the NIfTI files:

for t in eachindex(files)
    ni = niread(files[t])
    ni.raw .= img[:,:,:,t]
    niwrite(files[t], ni)
end

Note that this tutorial assumes that the headers of all NIfTI files are identical and replaces the raw data with interpolated data. For changing the NIfTI header instead, we can call params = realign!(img; mask=mask, realign=false) and write params to the NIfTI header. For more information, refer to the NIfTI.jl documentation.

Main Interface

MRIRealign.realign! — Function

realign!(img; center=size(img)[1:3] .÷ 2, ref_mode=:consensus, mask=trues(size(img)[1:3]), fwhm=nothing, realign=true)
realign!(img, motion_params; center=size(img)[1:3] .÷ 2)

Estimate motion parameters and/or realign the images.

Required argument

img::AbstractArray{T,4}: Array of images with the dimensions x, y, z, t, i.e., 3D images with time in the 4th dimension. The type T can be real or complex valued. If complex-valued, the motion estimation will be performed on the absolute value; otherwise, on img, i.e., allowing for negative values.

Keyword arguments if motion parameters are unknown

center=size(img)[1:3] .÷ 2: Point around which the images are rotated (relevant only for the estimated motion parameters, not the alignment).

ref_mode::consensus(default),:mean, or an integer.:consensusestimates motion parameters pairwise for all timeframes and computes a consensus, which is helpful if any single reference might have poor image quality. This comes at the cost of at-fold increase in computation time.:meanestimates the motion parameters wrt. to the mean of all images. This is fast, but might have inferior precision if the mean is blurred by substantial motion. Providing an integer aligns the images wrt. to theref_mode`th time frame, which is fast, but works only reliably if this time frame has good image quality.

mask=trues(size(img)[1:3]): bitmask at which the frames are compared.

fwhm=nothing: 3-Tuple of the full width at half maximum values of an optional Gaussian smoothing kernel along each dimension, in units of voxels. The default setting (nothing) is to perform no smoothing.

realign=true: If true, the argument img will be overwritten inline with the aligned images. If false, this function estimates the motion parameters, but does not align the images.

Optional arguments if the motion parameters are already known

motion_params::AbstractMatrix: If the motion parameters are known, e.g., by a previous run of this function, they can be provided to skip the estimation step. The dimensions of this matrix are 6 × T, capturing in the first dimension the motion parameters, in the order rx, ry, rz, tx, ty, tz, with the rotations r and the translations t. T is the number of time frames.

center=size(img)[1:3] .÷ 2): Point around which motion_params are applied.

With the appropriate settings (see above), the aligned timeframes are written inline into img. The function always returns the estimated motion parameters, where all rotations are in radians and translations in voxels.

Helper functions

MRIRealign.create_rotation_matrix — Function

create_rotation_matrix(rx, ry, rz)
create_rotation_matrix(p) = create_rotation_matrix(p[1], p[2], p[3])

Calculate the rotation matrix for three rotations rx, ry, rz, in radians. In this package, we use the convention R = Rz * Ry * Rx.

MRIRealign.create_affine_matrix — Function

create_affine_matrix(p, center)

Calculate the affine matrix from p = rx, ry, rz, tx, ty, tz. The rotations r are in radians and the translations t in voxels. The argument center takes a 3-Tuple (or vector of length 3) with the center around which the images are rotated.